+++ kitty/config_data.py
@@ -309,13 +309,15 @@ Choose the color of text under the cursor. If you want it rendered with the
 background color of the cell underneath instead, use the special keyword: background'''))
 o('cursor_shape', 'block', option_type=to_cursor_shape, long_text=_(
     'The cursor shape can be one of (block, beam, underline)'))
-o('cursor_blink_interval', 0.5, option_type=positive_float, long_text=_('''
+o('cursor_blink_interval', -1, option_type=float, long_text=_('''
 The interval (in seconds) at which to blink the cursor. Set to zero to disable
-blinking. Note that numbers smaller than :opt:`repaint_delay` will be limited
-to :opt:`repaint_delay`. Stop blinking cursor after the specified number of
-seconds of keyboard inactivity. Set to zero to never stop blinking.
+blinking. Negative values mean use system default. Note that numbers smaller
+than :opt:`repaint_delay` will be limited to :opt:`repaint_delay`.
 '''))
-o('cursor_stop_blinking_after', 15.0, option_type=positive_float)
+o('cursor_stop_blinking_after', 15.0, option_type=positive_float, long_text=_('''
+Stop blinking cursor after the specified number of seconds of keyboard
+inactivity.  Set to zero to never stop blinking.
+'''))
 
 # }}}